push 20-04-2023

+++ srcs/libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: emlamoth <emlamoth@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/06 16:29:06 by emlamoth          #+#    #+#             */
-/*   Updated: 2023/04/07 11:36:27 by emlamoth         ###   ########.fr       */
+/*   Updated: 2023/04/20 17:27:53 by emlamoth         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 /*The ft_atoi() function converts the initial portion of the string
 pointed to by <str> to int.*/
-long int	ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
-	int	i;
-	int	s;
+	int			i;
+	int			s;
 	long int	num;
 
 	i = 0;
